backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
nltk==3.8.1
textblob==0.17.1
newsapi-python==0.2.7
beautifulsoup4==4.12.2
lxml==4.9.3
gunicorn==21.2.0
schedule==1.2.0
pymongo==4.5.0
pycountry==22.3.5
feedparser==6.0.10
pandas==2.0.3
